requirements.txt
# Core pipeline dependencies (scripts/)
requests==2.32.3
beautifulsoup4==4.13.4
pdfplumber==0.11.9
xhtml2pdf==0.2.17
Markdown==3.10.2
notion-client==3.0.0
python-dotenv==1.2.1
playwright
PyMuPDF
python-docx
google-genai

# Agent (agent/)
anthropic
